+++ crates/release_plz_core/src/update.rs
@@ -3,7 +3,9 @@ use anyhow::{anyhow, Context};
 use cargo_metadata::{semver::Version, Package};
 use cargo_utils::upgrade_requirement;
 use cargo_utils::LocalManifest;
+use git_cmd::Repo;
 use std::{fs, path::Path};
+use tracing::info;
 
 use tracing::{debug, instrument};
 
@@ -23,7 +25,13 @@ pub fn update(input: &UpdateRequest) -> anyhow::Result<(Vec<(Package, UpdateResu
     if !packages_to_update.is_empty() {
         let local_manifest_dir = input.local_manifest_dir()?;
         update_cargo_lock(local_manifest_dir, input.should_update_dependencies())?;
+
+        let there_are_commits_to_push = Repo::new(local_manifest_dir)?.is_clean().is_err();
+        if !there_are_commits_to_push {
+            info!("the repository is already up-to-date");
+        }
     }
+
     Ok((packages_to_update, repository))
 }
 
@@ -74,8 +82,7 @@ fn set_version(
     local_manifest.set_package_version(version);
     local_manifest.write().expect("cannot update manifest");
 
-    let package_path =
-        fs::canonicalize(local_manifest.path.parent().context("at least a parent")?)?;
+    let package_path = fs::canonicalize(crate::manifest_dir(&local_manifest.path)?)?;
     update_dependencies(all_packages, version, &package_path)?;
     Ok(())
 }
@@ -88,11 +95,7 @@ fn update_dependencies(
 ) -> anyhow::Result<()> {
     for member in all_packages {
         let mut member_manifest = LocalManifest::try_new(member.manifest_path.as_std_path())?;
-        let member_dir = member_manifest
-            .path
-            .parent()
-            .context("at least a parent")?
-            .to_owned();
+        let member_dir = crate::manifest_dir(&member_manifest.path)?.to_owned();
         let deps_to_update = member_manifest
             .get_dependency_tables_mut()
             .flat_map(|t| t.iter_mut().filter_map(|(_, d)| d.as_table_like_mut()))